+++ app/models/vote.go
@@ -3,5 +3,5 @@ package models
 type Vote struct {
 	Nickname string `json:"nickname"`
 	Voice    int8   `json:"voice"`
-	Thread   int32  `json:"thread"`
+	Thread   int64  `json:"thread"`
 }
